dev server separeted from tests file
